bot/requirements.txt
requests
python-dotenv
pandas
openpyxl
router